+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Incorrect return types in Intel bulk APIs.
+
 ## [3.7.1] - 2024-03-20
 
 ### Added